requirements.txt
pypdf
python-docx
gethwp
pyhwp
ebooklib
lxml
markdown
pywebview
pyinstaller
